refactor: DefaultLogManager
