+++ p2p_wallet/Common/Services/Logger/LogManager.swift
@@ -1,13 +1,8 @@
+import AnalyticsManager
 import FeeRelayerSwift
 import Foundation
-import KeyAppBusiness
-import KeyAppKitCore
 import KeyAppKitLogger
-import Resolver
-import Send
 import SolanaSwift
-import UIKit
-import Wormhole
 
 enum LogLevel: String {
     case info
@@ -25,13 +20,7 @@ protocol LogManagerLogger {
     func log(event: String, logLevel: LogLevel, data: String?)
 }
 
-protocol LogManager {
-    func setProviders(_ providers: [LogManagerLogger])
-    func log(event: String, logLevel: LogLevel, data: String?, shouldLogEvent: (() -> Bool)?)
-    func log(event: String, logLevel: LogLevel, data: (any Encodable)?)
-}
-
-class DefaultLogManager: LogManager {
+class DefaultLogManager {
     static let shared: DefaultLogManager = {
         let manager = DefaultLogManager()
         SolanaSwift.Logger.setLoggers([manager])
@@ -49,8 +38,7 @@ class DefaultLogManager: LogManager {
         self.providers = providers
     }
 
-    func log(event: String, logLevel: LogLevel, data: String? = nil, shouldLogEvent: (() -> Bool)? = nil) {
-        guard shouldLogEvent?() ?? true else { return }
+    func log(event: String, logLevel: LogLevel, data: String? = nil) {
         providers.forEach { provider in
             guard provider.supportedLogLevels.contains(logLevel) else { return }
             queue.async {
@@ -60,23 +48,11 @@ class DefaultLogManager: LogManager {
     }
 
     func log(event: String, logLevel: LogLevel, data: (any Encodable)?) {
-        log(event: event, logLevel: logLevel, data: data?.jsonString, shouldLogEvent: nil)
-    }
-
-    func log(error: Error) {
-        // capture error
-        if let error = error as? CustomNSError {
-            log(event: "Error", logLevel: .error, data: error.errorUserInfo[NSDebugDescriptionErrorKey] as? String)
-        }
-        // else
-        else {
-            log(event: "Error", logLevel: .error, data: String(reflecting: error))
-        }
+        log(event: event, logLevel: logLevel, data: data?.jsonString)
     }
 }
 
-extension DefaultLogManager: SolanaSwiftLogger, FeeRelayerSwiftLogger, KeyAppKitLoggerType,
-KeyAppKitCore.ErrorObserver {
+extension DefaultLogManager: SolanaSwiftLogger, FeeRelayerSwiftLogger, KeyAppKitLoggerType {
     func log(event: String, data: String?, logLevel: LogLevel) {
         var newLogLevel: LogLevel = .info
         switch logLevel {
@@ -98,10 +74,7 @@ KeyAppKitCore.ErrorObserver {
             newLogLevel = .alert
         }
 
-        providers.forEach { logger in
-            guard logger.supportedLogLevels.contains(newLogLevel) else { return }
-            log(event: event, logLevel: newLogLevel, data: data)
-        }
+        log(event: event, logLevel: newLogLevel, data: data)
     }
 
     func log(event: String, data: String?, logLevel: SolanaSwift.SolanaSwiftLoggerLogLevel) {
@@ -148,99 +121,5 @@ KeyAppKitCore.ErrorObserver {
             newLogLevel = .debug
         }
         log(event: event, data: data, logLevel: newLogLevel)
-    }
-
-    func handleError(_ error: Error, config: ErrorObserverConfig?) {
-        let config = config ?? .init(flags: [])
-
-        // Default log
-        log(event: config.domain ?? "Error", logLevel: .error, data: String(reflecting: error))
-
-        // Realtime log
-        if config.flags.contains(.realtimeAlert) {
-            // Setup
-            let log: [String: Any] = [
-                "platform": "iOS \(UIDevice.current.systemVersion)",
-                "userPubkey": Resolver.resolve(UserWalletManager.self).wallet?.account.publicKey
-                    .base58EncodedString ?? "",
-                "appVersion": AppInfo.appVersionDetail,
-                "timestamp": "\(Int64(Date().timeIntervalSince1970 * 1000))",
-                "error": error.localizedDescription,
-            ]
-
-            let logStr: String
-            if
-                let logData = try? JSONSerialization.data(withJSONObject: log),
-                let log = String(data: logData, encoding: .utf8)
-            {
-                logStr = log
-            } else {
-                logStr = String(reflecting: log)
-            }
-
-            self.log(event: "\(config.domain ?? "General") iOS Alarm", logLevel: .alert, data: logStr)
-        }
-    }
-
-    func handleError(_ error: Error, userInfo: [String: Any]?) {
-        guard let error = error as? UserActionError else {
-            return
-        }
-
-        Task {
-            let data = await AlertLoggerDataBuilder.buildLoggerData(error: error)
-            let accountStorage: UserWalletManager = Resolver.resolve()
-            // Claim
-            if error.domain == WormholeClaimUserActionError.domain,
-               let action =
-               userInfo?[WormholeClaimUserActionError.UserInfoKey.action.rawValue] as? WormholeClaimUserAction
-            {
-                let message = ClaimAlertLoggerErrorMessage(
-                    tokenToClaim: .init(
-                        name: action.token.name,
-                        solanaMint: SupportedToken.ERC20(rawValue: action.token.erc20Address ?? "")?
-                            .solanaMintAddress ?? "",
-                        ethMint: action.token.id,
-                        claimAmount: CryptoFormatter().string(amount: action.amountInCrypto)
-                    ),
-                    userPubkey: data.userPubkey,
-                    userEthPubkey: accountStorage.wallet?.ethAddress ?? "",
-                    simulationError: error == WormholeClaimUserActionError.submitError ? error
-                        .localizedDescription : nil,
-                    bridgeSeviceError: error != WormholeClaimUserActionError.submitError ? error
-                        .localizedDescription : nil,
-                    blockchainError: nil
-                )
-                self.log(event: "Wormhole Claim iOS Alarm", logLevel: .alert, data: message)
-            } else if error.domain == WormholeSendUserActionError.domain,
-                      let action =
-                      userInfo?[WormholeClaimUserActionError.UserInfoKey.action.rawValue] as? WormholeSendUserAction
-            {
-                let simulationError: String? = {
-                    if error != WormholeSendUserActionError.feeRelaySignFailure && error != WormholeSendUserActionError
-                        .submittingToBlockchainFailure
-                    {
-                        return error.readableDescription
-                    }
-                    return nil
-                }()
-                let message = SendWormholeAlertLoggerErrorMessage(
-                    tokenToSend: .init(
-                        name: action.sourceToken.name,
-                        mint: action.sourceToken.id,
-                        sendAmount: CryptoFormatter().string(amount: action.amount)
-                    ),
-                    arbiterFeeAmount: action.fees.arbiter?.amount ?? "",
-                    userPubkey: data.userPubkey,
-                    recipientEthPubkey: action.recipient,
-                    simulationError: simulationError,
-                    feeRelayerError: error == WormholeSendUserActionError.feeRelaySignFailure ? error
-                        .readableDescription : nil,
-                    blockchainError: error == WormholeSendUserActionError.submittingToBlockchainFailure ? error
-                        .readableDescription : nil
-                )
-                self.log(event: "Wormhole Send iOS Alarm", logLevel: .alert, data: message)
-            }
-        }
     }
 }
